+++ pipeline.py
@@ -521,7 +521,7 @@ def run_in_amqp_mode(args):
 
     # Append tmp to output directory
     output_root = args.output
-    args.output = os.path.join(args.output, 'tmp')
+    args.output = os.path.join(args.output, args.model, 'tmp')
     os.makedirs(args.output, exist_ok=True)
 
     # Create Pipeline
@@ -551,6 +551,13 @@ def run_in_amqp_mode(args):
                     json_path = os.path.join(args.data, data['json_filename'])
                     map_name = os.path.splitext(os.path.basename(image_path))[0]
                     log.debug(f'RabbitMQ - {map_name} - Recieved cog')
+                    if map_name in active_maps:
+                        log.warning(f'RabbitMQ - {map_name} - Map already in pipeline, skipping')
+                        map_handle = {'method':method, 'properties':properties, 'data':data, 'id':data_id}
+                        map_handle['data']['exception'] = f'{map_name} - Map already in pipeline, skipping'
+                        channel.basic_publish(exchange='', routing_key=ERROR_QUEUE, body=json.dumps(map_handle['data']), properties=map_handle['properties'])
+                        channel.basic_ack(delivery_tag=map_handle['method'].delivery_tag)
+                        continue
                     active_maps[map_name] = {'method':method, 'properties':properties, 'data':data, 'id':data_id} # Keep track of maps we are working on.
                     input_data = (data['cog_id'], image_path, json_path)
                     input_stream.append(input_data)